+++ app.py
@@ -1,8 +1,12 @@
-# -----------------------------------------------------------------------------
-# Autor: Lic. Ricardo MONLA
-# Versión: v0.9.0
-# Descripción: rm-IAPromper - Backend Flask para generación de prompts estructurados con IA.
-# -----------------------------------------------------------------------------
+#  -----------------------------------------------------------------------------
+#  Project:     rm-Prompter
+#  File:        app.py
+#  Version:     v0.9.0
+#  Date:        2025-12-06
+#  Author:      Lic. Ricardo MONLA
+#  Email:       rmonla@gmail.com
+#  Description: rm-IAPromper - Backend Flask para generación de prompts estructurados con IA.
+#  -----------------------------------------------------------------------------
 
 from flask import Flask, jsonify, request
 import os