+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Todos los cambios notables en el proyecto **Asistente de Escritorio MFM** se doc
 
 El formato se basa en [Keep a Changelog](https://keepachangelog.com/es-ES/1.0.0/), y este proyecto se adhiere a [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## **[3.1.5] - 2025-12-06**
 
 ### **Añadido**